build: move FileAccess bindings rtti -> ast; tighten compiler lib
FileAccess is the compiler's source-file loader (opens .das paks, resolves
require paths, owns FileInfo cache). Nothing at pure runtime needs it.
Continues cff283e30 (which moved compile/compile_file/make_file_access out
of rtti) by collapsing the remaining FileAccess surface:

- FileAccess type annotation + introduceFile / rtti_add_file_access_root /
  rtti_add_extra_module bindings move from module_builtin_rtti.cpp ->
  module_builtin_ast.cpp. MAKE_EXTERNAL_TYPE_FACTORY moves rtti.h -> ast.h.
- daslib/rtti.das drops the FileAccessPtr typedef; daslib/ast.das gains it.
  API note: callers needing FileAccess now require daslib/ast (already the
  case for all in-tree callers via ast / ast_boost).
- FileAccess::serialize / ModuleFileAccess::serialize impls move out of
  module_builtin_ast_serialize.cpp into module_file_access.cpp where the
  classes live; an explicit AstSerializer::operator<<(das_hash_map<string,
  unique_ptr<FileInfo>>&) instantiation lives in serialize.cpp so the cross-
  TU template body resolves.

Also two side-effect cleanups uncovered during the link work:

- ast_derive_alias / ast_gc_report / ast_inscope_pod / ast_escape_analysis
  move from AST_SRC (both libs) to LIBDASCRIPT_COMPILER_SRC. These are
  compile-time analysis passes; nothing at runtime references them.
- module_builtin_ast_serialize.cpp moves to the end of BUILTIN_SRC (clearer
  group ordering: file_access right before its serialize uses).
- DAS_API tags on rtti_create_ast_serializer / _deserializer / _delete /
  serialize_program / deserialize_program / get_data so the runtime lib can
  reach them across the .so boundary.
- DAS_API on das::starts_with (called from compiler-side serialize.cpp).

Verified: full daslang link, 1042/1042 tests/language pass via dastest.
